add Dockerfile, refactor unittests
