+++ maunal_install.sh
