+++ requirements.txt
@@ -1,3 +1,3 @@
 fastapi==0.111.0
-uvicorn==0.30.0
-
+gunicorn==22.0.0
+pandas==2.2.2
