--- start.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-uvicorn app:app --reload
