+++ test.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 
-python3 -m coverage run -m unittest tests.py
+python3 -m coverage run -m unittest src/test_app.py
 python3 -m coverage html
